# docs/openapi.json
{
  "components": {
...
    "schemas": {
...
      "ActiveContext": {
        "additionalProperties": false,
        "properties": {
          "org_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "station_id"
        ],
        "type": "object"
      },
      "ActiveContextRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ActiveContextRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          }
        },
        "required": [
          "org_id",
          "station_id"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "description": "Stable machine error code",
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "field_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "message": {
            "description": "Safe user message",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/LoginRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "LossView": {
        "additionalProperties": false,
        "properties": {
          "cash_amount": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "liters": {
            "type": "string"
          },
          "loss_id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "row_id": {
            "type": "string"
          }
        },
        "required": [
          "row_id",
          "loss_id",
          "direction",
          "liters"
        ],
        "type": "object"
      },
...
      "ReadingView": {
        "additionalProperties": false,
        "properties": {
          "expected_sale": {
            "type": "string"
          },
          "meter_end": {
            "type": "string"
          },
          "meter_start": {
            "type": "string"
          },
          "nozzle_id": {
            "type": "string"
          }
        },
        "required": [
          "nozzle_id",
          "meter_start",
          "meter_end",
          "expected_sale"
        ],
        "type": "object"
      },
      "ReportView": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ReportView.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "losses": {
            "items": {
              "$ref": "#/components/schemas/LossView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/ReadingView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "report_id": {
            "type": "string"
          },
          "sales": {
            "items": {
              "$ref": "#/components/schemas/SalesView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "shift_id": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "submitted_at": {
            "type": "string"
          },
          "version_no": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "station_id",
          "shift_id",
          "version_no",
          "status",
          "submitted_at",
          "readings",
          "sales",
          "losses"
        ],
        "type": "object"
      },
...
      "SalesView": {
        "additionalProperties": false,
        "properties": {
          "cash_amount": {
            "type": "string"
          },
          "cashless_amount": {
            "type": "string"
          },
          "dispenser_id": {
            "type": "string"
          }
        },
        "required": [
          "dispenser_id",
          "cash_amount",
          "cashless_amount"
        ],
        "type": "object"
      },
      "SessionView": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/SessionView.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "active_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "station_ids": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "roles",
          "org_id",
          "station_ids",
          "active_context"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "httpOnly session cookie",
        "in": "cookie",
        "name": "pomkita_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "PomKita API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example": {
                  "summary": "Request example",
                  "value": {
                    "password": "demo-password",
                    "username": "demo-owner"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create a session",
        "x-csrf-required": true,
        "x-stable-error-codes": {
          "401": [
            "invalid_credentials"
          ]
        }
      }
    },
    "/api/v1/logout": {
      "delete": {
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke a session"
      }
    },
...
    "/api/v1/reports/{id}": {
      "get": {
        "operationId": "getReport",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "station_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read a report"
      }
    },
...
    "/api/v1/reports/{id}/printout": {
      "get": {
        "operationId": "printReport",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "station_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read report printout"
      }
    },
    "/api/v1/session": {
      "get": {
        "operationId": "session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read the current session",
        "x-permitted-roles": [
          "Operator",
          "Supervisor",
          "Station Admin",
          "Owner",
          "Superadmin"
        ]
      }
    },
    "/api/v1/session/active-context": {
      "post": {
        "operationId": "setActiveContext",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example": {
                  "summary": "Request example",
                  "value": {
                    "org_id": "11111111-1111-4111-8111-111111111111",
                    "station_id": "22222222-2222-4222-8222-222222222222"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ActiveContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Set the active organization and station for the current session",
        "x-csrf-required": true,
        "x-permitted-roles": [
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "400": [
            "validation_error"
          ],
          "401": [
            "invalid_session"
          ],
          "403": [
            "active_context_forbidden",
            "csrf_required"
          ],
          "404": [
            "active_context_target_not_found"
          ]
        }
      }
    },
...
  },
...
}