# internal/httpapi/session/handlers.go
// Package session provides session HTTP handlers.
package session

import (
	"context"
	"errors"
	"net/http"
	"strings"

	"github.com/fadhln/pomkita-be/internal/httpapi/transport"
	appjwt "github.com/fadhln/pomkita-be/internal/jwt"
	appauth "github.com/fadhln/pomkita-be/internal/service/auth"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

// Service provides the session use cases required by HTTP handlers.
type Service interface {
	Login(context.Context, string, string) (string, appjwt.Claims, error)
	Logout(context.Context, uuid.UUID) error
	ReadSession(context.Context, string) (appjwt.SessionView, error)
	SetActiveContext(context.Context, uuid.UUID, []string, uuid.UUID, uuid.UUID) error
}

// LoginRequest contains session login credentials.
type LoginRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

// ActiveContextRequest selects the active organization and station for a session.
type ActiveContextRequest struct {
	OrgID     uuid.UUID `json:"org_id"`
	StationID uuid.UUID `json:"station_id"`
}

// RegisterRoutes registers login, logout, and session routes.
func RegisterRoutes(router gin.IRoutes, verifier transport.TokenVerifier, service Service, secure bool) {
	router.POST("/login", transport.RequireCSRF, loginHandler(service, secure))
	router.DELETE("/logout", transport.AuthMiddleware(verifier), transport.RequireCSRF, logoutHandler(service, secure))
	router.GET("/session", transport.AuthMiddleware(verifier), sessionHandler(service))
	router.POST("/session/active-context", transport.AuthMiddleware(verifier), transport.RequireCSRF, activeContextHandler(service))
}

func loginHandler(service Service, secure bool) gin.HandlerFunc {
	return func(c *gin.Context) {
		if service == nil {
			transport.WriteError(c, http.StatusInternalServerError, "internal_error")
			return
		}
		var input LoginRequest
		if !transport.DecodeRequest(c, &input) {
			return
		}
		if strings.TrimSpace(input.Username) == "" || input.Password == "" {
			transport.ValidationError(c)
			return
		}
		token, _, err := service.Login(c.Request.Context(), strings.ToLower(strings.TrimSpace(input.Username)), input.Password)
		if err != nil {
			if errors.Is(err, appauth.ErrInvalidCredentials) {
				transport.WriteError(c, http.StatusUnauthorized, "invalid_credentials")
				return
			}
			_ = c.Error(err)
			return
		}
		c.SetSameSite(http.SameSiteLaxMode)
		c.SetCookie("pomkita_session", token, 900, "/", "", secure, true)
		c.Status(http.StatusOK)
	}
}

func logoutHandler(service Service, secure bool) gin.HandlerFunc {
	return func(c *gin.Context) {
		if service == nil {
			transport.WriteError(c, http.StatusInternalServerError, "internal_error")
			return
		}
		claims := c.MustGet("jwt_claims").(appjwt.Claims)
		if err := service.Logout(c.Request.Context(), claims.JTI); err != nil {
			_ = c.Error(err)
			return
		}
		c.SetSameSite(http.SameSiteLaxMode)
		c.SetCookie("pomkita_session", "", -1, "/", "", secure, true)
		c.Status(http.StatusNoContent)
	}
}

func activeContextHandler(service Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		if service == nil {
			transport.WriteError(c, http.StatusInternalServerError, "internal_error")
			return
		}
		var input ActiveContextRequest
		if !transport.DecodeRequest(c, &input) {
			return
		}
		if input.OrgID == uuid.Nil || input.StationID == uuid.Nil {
			transport.ValidationError(c)
			return
		}
		view, err := service.ReadSession(c.Request.Context(), c.GetString("raw_token"))
		if err != nil {
			_ = c.Error(err)
			return
		}
		claims := c.MustGet("jwt_claims").(appjwt.Claims)
		if err := service.SetActiveContext(c.Request.Context(), claims.JTI, view.Roles, input.OrgID, input.StationID); err != nil {
			_ = c.Error(err)
			return
		}
		c.Status(http.StatusNoContent)
	}
}

func sessionHandler(service Service) gin.HandlerFunc {
	return func(c *gin.Context) {
		if service == nil {
			transport.WriteError(c, http.StatusInternalServerError, "internal_error")
			return
		}
		view, err := service.ReadSession(c.Request.Context(), c.GetString("raw_token"))
		if err != nil {
			_ = c.Error(err)
			return
		}
		c.JSON(http.StatusOK, view)
	}
}